Mise en place Formulaire Participant Mise en place préchargement participant

+++ src/com/cptbloc/dao/CategorieDAO.java
@@ -4,5 +4,5 @@ import com.cptbloc.beans.Categorie;
 
 public interface CategorieDAO {
 
-    Categorie trouverAgeCategorie( Long idDefCategorie ) throws DAOException;
+    Categorie trouverAgeCategorie( int idDefCategorie ) throws DAOException;
 }
